reordered the settings screens
