--- app/src/main/res/xml/prefs_screen_advanced.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright (C) 2014 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
-
-<PreferenceScreen
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:latin="http://schemas.android.com/apk/res-auto"
-    android:title="@string/settings_screen_advanced"
-    android:key="screen_advanced">
-    <rkr.simplekeyboard.inputmethod.latin.settings.SeekBarDialogPreference
-        android:key="pref_vibration_duration_settings"
-        android:title="@string/prefs_keypress_vibration_duration_settings"
-        latin:maxValue="@integer/config_max_vibration_duration" />
-    <rkr.simplekeyboard.inputmethod.latin.settings.SeekBarDialogPreference
-        android:key="pref_keypress_sound_volume"
-        android:title="@string/prefs_keypress_sound_volume_settings"
-        latin:maxValue="100" /> <!-- percent -->
-    <rkr.simplekeyboard.inputmethod.latin.settings.SeekBarDialogPreference
-        android:key="pref_key_longpress_timeout"
-        android:title="@string/prefs_key_longpress_timeout_settings"
-        latin:minValue="@integer/config_min_longpress_timeout"
-        latin:maxValue="@integer/config_max_longpress_timeout"
-        latin:stepValue="@integer/config_longpress_timeout_step" />
-    <rkr.simplekeyboard.inputmethod.latin.settings.SeekBarDialogPreference
-        android:key="pref_keyboard_height"
-        android:title="@string/prefs_keyboard_height_settings"
-        latin:minValue="@integer/config_min_keyboar_height"
-        latin:maxValue="@integer/config_max_keyboar_height"
-        latin:stepValue="@integer/config_keyboar_height_step" />
-    <rkr.simplekeyboard.inputmethod.latin.settings.ColorDialogPreference
-        android:key="pref_keyboard_color"
-        android:title="@string/keyboard_color" />
-    <CheckBoxPreference
-        android:key="pref_matching_navbar_color"
-        android:title="@string/matching_navbar_color"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_hide_special_chars"
-        android:title="@string/hide_special_chars"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_hide_language_switch_key"
-        android:title="@string/hide_language_switch_key"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_enable_ime_switch"
-        android:title="@string/pref_enable_ime_switch"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_show_number_row"
-        android:title="@string/show_number_row"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_space_swipe"
-        android:title="@string/space_swipe"
-        android:defaultValue="false" />
-    <CheckBoxPreference
-        android:key="pref_delete_swipe"
-        android:title="@string/delete_swipe"
-        android:defaultValue="false" />
-</PreferenceScreen>
