+++ app/src/main/res/xml/prefs_screen_appearance.xml
@@ -16,14 +16,24 @@
 
 <PreferenceScreen
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:latin="http://schemas.android.com/apk/res-auto"
     android:key="screen_appearance"
     android:title="@string/settings_screen_appearance">
     <PreferenceScreen
         android:fragment="rkr.simplekeyboard.inputmethod.latin.settings.ThemeSettingsFragment"
         android:key="screen_theme"
         android:title="@string/settings_screen_theme" />
-    <PreferenceScreen
-        android:fragment="rkr.simplekeyboard.inputmethod.latin.settings.CustomInputStyleSettingsFragment"
-        android:key="custom_input_styles"
-        android:title="@string/custom_input_styles_title" />
+    <rkr.simplekeyboard.inputmethod.latin.settings.ColorDialogPreference
+        android:key="pref_keyboard_color"
+        android:title="@string/keyboard_color" />
+    <CheckBoxPreference
+        android:key="pref_matching_navbar_color"
+        android:title="@string/matching_navbar_color"
+        android:defaultValue="false" />
+    <rkr.simplekeyboard.inputmethod.latin.settings.SeekBarDialogPreference
+        android:key="pref_keyboard_height"
+        android:title="@string/prefs_keyboard_height_settings"
+        latin:minValue="@integer/config_min_keyboar_height"
+        latin:maxValue="@integer/config_max_keyboar_height"
+        latin:stepValue="@integer/config_keyboar_height_step" />
 </PreferenceScreen>
